[WIP] Move adjuster to a separate file
